feat: enabling instnaces and moving to new component
